tests/requirements/base.txt
django-filer
easy-thumbnails
djangocms-text
djangocms-link
coverage
django-app-helper
django-treebeard<4.5
tox
coverage
isort
flake8
pyflakes>=2.1
wheel
black
pre-commit
svglib!=1.6.0
